requirements.txt
# ============================================
# PIKCHA ETL - Python Dependencies
# ============================================

# --------------------------------------------
# Базы данных
# --------------------------------------------
pymongo==4.16.0              # MongoDB driver
clickhouse-driver==0.2.10    # ClickHouse client

# --------------------------------------------
# Стриминг данных
# --------------------------------------------
kafka-python==2.3.0          # Kafka client (Producer/Consumer)

# --------------------------------------------
# Генерация данных
# --------------------------------------------
Faker==40.4.0                # Генерация синтетических данных

# --------------------------------------------
# Работа с S3
# --------------------------------------------
boto3==1.42.54               # AWS SDK для S3

# --------------------------------------------
# Утилиты и конфигурация
# --------------------------------------------
python-dotenv==1.2.1         # Управление переменными окружения (.env)

